ml/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
tensorflow==2.13.0
torch==2.0.1
joblib==1.3.2
python-dotenv==1.0.0
requests==2.31.0
psycopg2-binary==2.9.7
redis==4.6.0
schedule==1.2.0
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0
gunicorn==21.2.0
